Refactor backend and clean unused components
